Improve support for exceptions thrown by callbacks when the same 'dll' function calls multiple callbacks in sequence before returning.
In this commit, I added an exception check at the first possible opportunity at the start of the C++ callback implementations. This is to ensure that a callback doesn't execute at all if a JNI exception was raised before it was called.

How could such an exception state possibly happen?????

At the moment, it can happen if the same invocation of a 'dll' function calls multiple callbacks. Recall that a 'dll' function is a black box, usually a Win32 API function written in plain C and/or assembly language. The 'dll' has no idea whether or not a callback raised a JNI exception or not. So imagine a 'dll' function implemented like this:

    __stdcall void dll_func(callback1_ptr f, calblack2_ptr g)
    {
        ...  // do some stuff
        f(); // call first callback
        ...  // do some more stuff
        g(); // call second callback
    }

Now if f() causes a Java/JNI exception to be raised, JNI will be in an exception state when g() is called. this exception check
at least ensure that g() will not run.

THIS IS A STOPGAP SOLUTION! The appropriate solution is to:

1) Wrap the callback invocation in a Win32 SEH __try/__except block.
2) Ensure that any callback which raises a JNI exception
    (i)  cleans up after itself; and then
    (ii) raises an SEH exception which should cause SEH to immediately unwind out to the next SEH __except block.
3) Wrap stdcall_invoke calls in SEH as well.

However, because Microsoft's cl.exe is the only compiler package which supports SEH, we will need to build exclusively in VS to take that step.